requirements.txt
Django
djangorestframework
Pillow
Faker
requests
django-cors-headers